feat: remove files && add tab animation
